DATAMONGO-1348 - Convert GeoJson used in NearQuery.
We now convert GeoJsonPoint used in NearQuery into its according format. This also requires to convert any given min/maxDistance as well as the distanceMultiplier into meters (metric system).

Along the way we fixed an issue where the actual Query used along with NearQuery was not properly mapped to the domain types properties.